+++ apps/meteor/client/methods/index.ts
@@ -1,5 +1,0 @@
-import './hideRoom';
-import './openRoom';
-import './pinMessage';
-import './unpinMessage';
-import './updateMessage';