+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -1,7 +1,0 @@
-import './jumpToMessage';
-import './jumpToPinMessage';
-import './jumpToSearchMessage';
-import './jumpToStarMessage';
-import './permalinkPinned';
-import './pinMessage';
-import './unpinMessage';