+++ packages/cron/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rocket.chat/cron
 
+## 0.1.10-rc.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.1.0-rc.1
+  - @rocket.chat/models@1.0.1-rc.1
+  </details>
+
+## 0.1.10-rc.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies [80e36bfc3938775eb26aa5576f1b9b98896e1cc4, 32d93a0666fa1cbe857d02889e93d9bbf45bd4f0]:</summary>
+
+  - @rocket.chat/core-typings@7.1.0-rc.0
+  - @rocket.chat/models@1.0.1-rc.0
+  </details>
+
 ## 0.1.9
 
 ### Patch Changes